+++ code/src/sixsq/nuvla/server/resources/nuvlabox/workflow_utils.clj
@@ -573,9 +573,10 @@
     (doseq [id ids]
       (let [request {:params      {:uuid          (u/id->uuid id)
                                    :resource-name nb-peripheral/resource-type}
-                     :body        {:acl (utils/set-acl-nuvlabox-view-only
-                                          nuvlabox-acl
-                                          {:owners [nuvlabox-id]})}
+                     :body        {:acl (-> nuvlabox-acl
+                                            (utils/set-acl-nuvlabox-view-only
+                                              {:owners [nuvlabox-id]})
+                                            (assoc :manage (:view-acl nuvlabox-acl)))}
                      :nuvla/authn auth/internal-identity}
             {status :status :as resp} (crud/edit request)]
         (if (= 200 status)