+++ code/src/sixsq/nuvla/server/resources/nuvlabox_peripheral.clj
@@ -4,12 +4,17 @@ The nuvlabox-peripheral resource represents a peripheral attached to a
 nuvlabox.
 "
   (:require
+    [sixsq.nuvla.auth.acl-resource :as a]
+    [sixsq.nuvla.auth.utils :as auth]
+    [sixsq.nuvla.db.impl :as db]
     [sixsq.nuvla.server.resources.common.crud :as crud]
     [sixsq.nuvla.server.resources.common.std-crud :as std-crud]
     [sixsq.nuvla.server.resources.common.utils :as u]
+    [sixsq.nuvla.server.resources.job :as job]
     [sixsq.nuvla.server.resources.nuvlabox.utils :as utils]
     [sixsq.nuvla.server.resources.resource-metadata :as md]
     [sixsq.nuvla.server.resources.spec.nuvlabox-peripheral :as nb-peripheral]
+    [sixsq.nuvla.server.util.log :as logu]
     [sixsq.nuvla.server.util.metadata :as gen-md]
     [sixsq.nuvla.server.util.response :as r]))
 
@@ -23,6 +28,77 @@ nuvlabox.
 (def collection-acl {:add   ["group/nuvla-user"]
                      :query ["group/nuvla-user"]})
 
+
+(defn has-video-capability?
+  [resource]
+  (contains? resource :video-device))
+
+
+(defn create-job
+  [{:keys [id parent]} request action]
+  (try
+    (let [user-id (auth/current-user-id request)
+          {{job-id     :resource-id
+            job-status :status} :body} (job/create-job id action
+                                                       {:owners   ["group/nuvla-admin"]
+                                                        :edit-acl [user-id]}
+                                                       :priority 50
+                                                       :affected-resources [{:href id}
+                                                                            {:href parent}])
+          job-msg (str "starting " id " with async " job-id)]
+      (when (not= job-status 201)
+        (throw (r/ex-response
+                 (format "unable to create async job to %s" action) 500 id)))
+      (r/map-response job-msg 202 id job-id))
+    (catch Exception e
+      (or (ex-data e) (throw e)))))
+
+
+(defn throw-data-gateway-already-enabled
+  [{:keys [data-gateway-enabled] :as resource}]
+  (if data-gateway-enabled
+    (logu/log-and-throw-400 "NuvlaBox peripheral data gateway already enabled!")
+    resource))
+
+
+(defn throw-data-gateway-already-disabled
+  [{:keys [data-gateway-enabled] :as resource}]
+  (if data-gateway-enabled
+    resource
+    (logu/log-and-throw-400 "NuvlaBox peripheral data gateway already disabled!")))
+
+
+(defn throw-doesnt-have-video-capability
+  [resource]
+  (if (has-video-capability? resource)
+    resource
+    (logu/log-and-throw-400 "NuvlaBox peripheral does not have video capability!")))
+
+
+(defmethod crud/do-action [resource-type "enable-stream"]
+  [{{uuid :uuid} :params :as request}]
+  (try
+    (let [id (str resource-type "/" uuid)]
+      (-> (db/retrieve id request)
+          (a/throw-cannot-manage request)
+          (throw-doesnt-have-video-capability)
+          (throw-data-gateway-already-enabled)
+          (create-job request "enable-stream")))
+    (catch Exception e
+      (or (ex-data e) (throw e)))))
+
+
+(defmethod crud/do-action [resource-type "disable-stream"]
+  [{{uuid :uuid} :params :as request}]
+  (try
+    (let [id (str resource-type "/" uuid)]
+      (-> (db/retrieve id request)
+          (a/throw-cannot-manage request)
+          (throw-data-gateway-already-disabled)
+          (throw-doesnt-have-video-capability)
+          (create-job request "disable-stream")))
+    (catch Exception e
+      (or (ex-data e) (throw e)))))
 
 ;;
 ;; multimethods for validation
@@ -53,9 +129,10 @@ nuvlabox.
   [resource request]
   (when-let [nuvlabox-id (:parent resource)]
     (let [{nuvlabox-acl :acl} (crud/retrieve-by-id-as-admin nuvlabox-id)]
-      (assoc resource :acl (utils/set-acl-nuvlabox-view-only
-                             nuvlabox-acl
-                             {:owners [nuvlabox-id]})))))
+      (assoc resource :acl
+                      (-> nuvlabox-acl
+                          (utils/set-acl-nuvlabox-view-only {:owners [nuvlabox-id]})
+                          (assoc :manage (:view-acl nuvlabox-acl)))))))
 
 
 ;;
@@ -90,7 +167,15 @@ nuvlabox.
 
 
 (defmethod crud/delete resource-type
-  [request]
+  [{{uuid :uuid} :params :as request}]
+  (try
+    (let [id (str resource-type "/" uuid)
+          {:keys [data-gateway-enabled] :as resource} (db/retrieve id request)]
+
+      (when (and data-gateway-enabled (has-video-capability? resource))
+        (create-job resource request "disable-stream")))
+    (catch Exception e
+      (or (ex-data e) (throw e))))
   (delete-impl request))
 
 
@@ -100,6 +185,25 @@ nuvlabox.
 (defmethod crud/query resource-type
   [request]
   (query-impl request))
+
+
+;;
+;; Set operation
+;;
+
+
+
+(defmethod crud/set-operations resource-type
+  [{:keys [id data-gateway-enabled] :as resource} request]
+  (let [enable-stream-op  (u/action-map id :enable-stream)
+        disable-stream-op (u/action-map id :disable-stream)
+        can-manage?       (a/can-manage? resource request)
+        has-video?        (has-video-capability? resource)]
+    (cond-> (crud/set-standard-operations resource request)
+            (and can-manage?
+                 has-video?) (update :operations conj (if data-gateway-enabled
+                                                        disable-stream-op
+                                                        enable-stream-op)))))
 
 
 ;;